requirements.txt
akismet==1.1
boto3==1.12.35
botocore==1.15.35
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-contact-form==1.8.1
django-crispy-forms==1.9.0
django-environ==0.4.5
django-mail-admin==0.1.4
django-storages==1.9.1
docutils==0.15.2
idna==2.9
jmespath==0.9.5
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
